solved issue in ubuntu

+++ applications/GeoMechanicsApplication/custom_elements/small_strain_U_Pw_diff_order_element.cpp
@@ -508,10 +508,10 @@ void SmallStrainUPwDiffOrderElement::CalculateOnIntegrationPoints(const Variable
             << r_geom.WorkingSpaceDimension() << " in element: " << this->Id() << std::endl;
         size_t variable_index = 0;
         if (rVariable == CONFINED_STIFFNESS) {
-            variable_index = r_geom.WorkingSpaceDimension() == 2 ? INDEX_2D_PLANE_STRAIN_XX : INDEX_3D_XX;
+            variable_index = r_geom.WorkingSpaceDimension() == 2 ? static_cast<size_t>(INDEX_2D_PLANE_STRAIN_XX) : static_cast<size_t>(INDEX_3D_XX);
         } else {
             // rVariable == SHEAR_STIFFNESS
-            variable_index = r_geom.WorkingSpaceDimension() == 2 ? INDEX_2D_PLANE_STRAIN_XY : INDEX_3D_XZ;
+            variable_index = r_geom.WorkingSpaceDimension() == 2 ? static_cast<size_t>(INDEX_2D_PLANE_STRAIN_XY) : static_cast<size_t>(INDEX_3D_XZ);
         }
 
         ElementVariables Variables;
